Generate Rarefaction Curves

This function generates rarefaction curves for visualizing species richness at different sequencing depths, based on the OTU table in a phyloseq object.

Usage

amp_rarecurve(
  data,
  step = 100,
  ylim = NULL,
  xlim = NULL,
  label = FALSE,
  color = NULL,
  rank = rank,
  legend = TRUE,
  color.vector = NULL,
  legend.position = "topleft"
)

Arguments

data

A phyloseq object containing OTU data and associated metadata.

step

An integer specifying the step size for the rarefaction curve. Default is 100.

ylim

A numeric vector of length two specifying the y-axis limits. Default is NULL, which automatically adjusts the y-axis.

xlim

A numeric vector of length two specifying the x-axis limits. Default is NULL, which automatically adjusts the x-axis.

label

A logical value indicating whether sample labels should be displayed on the plot. Default is FALSE.

color

A character string specifying the column in the sample metadata to use for coloring the samples. Default is NULL (no color applied).

rank

An optional rank level for taxa aggregation. Default is rank.

legend

A logical value indicating whether to display a legend on the plot. Default is TRUE.

color.vector

An optional character vector specifying custom colors for groups in the color column. Default is NULL (default colors are used).

legend.position

A character string specifying the position of the legend. Default is "topleft".

Value

A rarefaction curve plot.

Examples

amp_rarecurve(data = physeq, step = 200, color = "sample_type", legend.position = "topright")
#> Error in amp_rarecurve(data = physeq, step = 200, color = "sample_type",     legend.position = "topright"): could not find function "amp_rarecurve"